requirements.txt
fastapi
python-multipart
google-generativeai
python-dotenv
hypercorn
pydantic